+++ math/linear_math/linear_math_test.cpp
@@ -2,13 +2,15 @@
 #include "math\linear_math\linear_math.h"
 #include "math\general_math\general_math.h"
 
+
 void IsLinearTest() {
     IsLinear_tester test[] ={{ 1, 0},
                              {10, 0},
                              {-1, 0},
                              { 0, 1}};
-    unsigned int test_amount = sizeof(test)/sizeof(test[0]);
-    for (unsigned int i = 0; i < test_amount; i++) {
+
+    size_t test_amount = sizeof(test)/sizeof(test[0]);
+    for (size_t i = 0; i < test_amount; i++) {
         IsLinear_tester recent = test[i];
         bool result = IsLinear(recent.a);
         if (result != recent.expected) {
@@ -23,8 +25,9 @@ void RootsOfLinearTest() {
     RootsOfLinear_tester test[] ={{{0,  2,  3, NAN, NAN, ZeroSolutions}, -1.5, -1.5},
                                   {{0,  2,  0, NAN, NAN, ZeroSolutions},    0,    0},
                                   {{0,  2, -4, NAN, NAN, ZeroSolutions},    2,    2}};
-    unsigned int test_amount = sizeof(test)/sizeof(test[0]);
-    for (unsigned int i = 0; i < test_amount; i++) {
+
+    size_t test_amount = sizeof(test)/sizeof(test[0]);
+    for (size_t i = 0; i < test_amount; i++) {
         RootsOfLinear(&test[i].eq);
         if (CompareWithZero(test[i].eq.x1-test[i].x1_expected) != Equals || CompareWithZero(test[i].eq.x2-test[i].x2_expected) != Equals) {
             printf("FAILED: test #%d\nexpected -> x1 = %lf, x2 = %lf\nanswer -> x1 = %lf, x2 = %lf\n\n",
@@ -32,3 +35,4 @@ void RootsOfLinearTest() {
         }
     }
 }
+
